header cleanup
As functionality moved around I didn't keep up with cleaning up the headers.
Try to clean up the headers so that the headers included in the various files
are ones that are actually needed for the function calls being made, objects
being referenced, and types actually being used in the file itself.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>